Update json gem
I was getting local failures, but travis seems to have been fine? But I've been
running a lot of work and specs locally with this newer version of json, so it
seems safe.

We DO have to update the bundler version that runs on travis, though, to match
Gemfile.lock. We won't have to do this for more modern rubies, but for 2.5.* we
still do.